Reduce visibility of ListenerContainerImpl classes, make ListenerDelegate internal API
